% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_.R
\name{boxcox_reverse}
\alias{boxcox_reverse}
\title{Box-Cox 逆变换函数}
\usage{
boxcox_reverse(transformed_data, lambda)
}
\arguments{
\item{transformed_data}{Box-Cox 转换后的数据}

\item{lambda}{Box-Cox 转换使用的 lambda 参数}
}
\value{
还原后的原始数据
}
\description{
将 Box-Cox 转换后的数据还原为原始尺度
}
\examples{
original <- c(1, 5, 10, 20, 50, 100)
result <- maxlikelihood_boxcox(original, return_lambda = TRUE)
reversed <- boxcox_reverse(result$transformed_data, result$lambda)
}
